requirements.txt
spacy==2.3.5
medspacy==0.0.1.4
Pillow==8.0.1
en-info-3700-i2b2-2012 @ https://github.com/abchapman93/spacy_models/raw/master/releases/en_info_3700_i2b2_2012-0.1.0/dist/en_info_3700_i2b2_2012-0.1.0.tar.gz